requirements.txt
streamlit
pandas
joblib
requests
plotly
matplotlib
seaborn
python-dotenv
xgboost
pyarrow
numpy
scikit-learn
folium<0.21,>=0.18.3 

